+++ src/test/java/com/api/record/model/CreateJobPayload.java
@@ -1,4 +1,4 @@
-package com.api.pojo;
+package com.api.record.model;
 
 import java.util.List;
