+++ src/test/java/com/api/record/model/CustomerProduct.java
@@ -1,4 +1,4 @@
-package com.api.pojo;
+package com.api.record.model;
 
 public record CustomerProduct(String dop, String serial_number, String imei1, String imei2, String popurl,
 		int product_id, int mst_model_id) {
