+++ src/test/java/com/api/record/model/Problems.java
@@ -1,4 +1,4 @@
-package com.api.pojo;
+package com.api.record.model;
 
 public record Problems(int id, String remark) {
