+++ src/test/java/com/api/record/model/UserCredentials.java
@@ -1,4 +1,4 @@
-package com.api.pojo;
+package com.api.record.model;
 
 public record UserCredentials(String username,String password) {
